move method GetDetph to base class since used in AliHLTEMCALGeometry and call in this class AliEMCALRecoUtilsBase
