+++ EMCAL/EMCALUtils/AliEMCALRecoUtilsBase.cxx
@@ -13,6 +13,10 @@
  * provided "as is" without express or implied warranty.                  *
  **************************************************************************/
 
+// ROOT includes
+#include <TGeoManager.h>
+#include <TGeoBBox.h>
+
 // STEER includes
 #include "AliVCluster.h"
 #include "AliESDtrack.h"
@@ -317,4 +321,76 @@ Bool_t AliEMCALRecoUtilsBase::ExtrapolateTrackToCluster(AliExternalTrackParam *t
   return ExtrapolateTrackToPosition(trkParam, clsPos, mass, step, tmpEta, tmpPhi);
 }
 
+///
+/// Calculate shower depth for a given cluster energy and particle type.
+/// Needed to calculate cluster position.
+///
+/// \param energy: cluster energy
+/// \param iParticle: particle assumption defined in enum ParticleType
+/// \param iSM: supermodule number
+///
+//_________________________________________________________
+Float_t  AliEMCALRecoUtilsBase::GetDepth(Float_t energy, 
+                                         Int_t   iParticle, 
+                                         Int_t   iSM ) const 
+{
+  // parameters 
+  Float_t x0    = 1.31;
+  Float_t ecr   = 8;
+  Float_t depth = 0;
+  Float_t arg   = energy*1000/ ecr; //Multiply energy by 1000 to transform to MeV
+  
+  switch ( iParticle )
+  {
+    case kPhoton:
+      if (arg < 1) 
+        depth = 0;
+      else
+        depth = x0 * (TMath::Log(arg) + 0.5); 
+      break;
+      
+    case kElectron:
+      if (arg < 1) 
+        depth = 0;
+      else
+        depth = x0 * (TMath::Log(arg) - 0.5); 
+      break;
+      
+    case kHadron:
+      // hadron 
+      // boxes anc. here
+      if (gGeoManager) 
+      {
+        gGeoManager->cd("ALIC_1/XEN1_1");
+        TGeoNode        *geoXEn1    = gGeoManager->GetCurrentNode();
+        TGeoNodeMatrix  *geoSM      = dynamic_cast<TGeoNodeMatrix *>(geoXEn1->GetDaughter(iSM));
+        if (geoSM) 
+        {
+          TGeoVolume      *geoSMVol   = geoSM->GetVolume(); 
+          TGeoShape       *geoSMShape = geoSMVol->GetShape();
+          TGeoBBox        *geoBox     = dynamic_cast<TGeoBBox *>(geoSMShape);
+          if (geoBox) depth = 0.5 * geoBox->GetDX()*2 ;
+          else AliFatal("Null GEANT box");
+        }
+        else AliFatal("NULL  GEANT node matrix");
+      }
+      else
+      {//electron
+        if (arg < 1) 
+          depth = 0;
+        else
+          depth = x0 * (TMath::Log(arg) - 0.5); 
+      }
+      
+      break;
+      
+    default://photon
+      if (arg < 1) 
+        depth = 0;
+      else
+        depth = x0 * (TMath::Log(arg) + 0.5);
+  }  
+  
+  return depth;
+}
